+++ test/test_helper.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
+$LOAD_PATH.unshift File.expand_path('../../lib', __dir__)
 require 'canql'
 
 require 'minitest/autorun'
@@ -37,6 +37,7 @@ end
 
 def assert_dir_block_count(parser, block_type, numder_of_blocks)
   return true if parser.meta_data[block_type].blank? && numder_of_anomalies.zero?
+
   assert parser.meta_data[block_type].present?, 'No DIR blocks to count'
   assert_equal numder_of_blocks, parser.meta_data.dig(block_type, Canql::ALL).count
 end
@@ -58,12 +59,12 @@ end
 
 def assert_dir_block_keys_valid(dir_block, all_keys = nil)
   assert dir_block.keys.reject { |key| all_keys.include?(key) }.none?,
-         'DIR block has unknown keys'
+         "DIR block has unknown keys #{dir_block.keys.reject { |key| all_keys.include?(key) }}"
 end
 
 def assert_dir_block_has_expected_keys(dir_block, expected = {})
-  assert expected.keys.reject { |key| dir_block.include?(key) }.none?,
-         'Expected key not present in DIR block'
+  missing_keys = expected.keys.reject { |key| dir_block.include?(key) }
+  assert missing_keys.none?, "Expected key not present in DIR block: #{missing_keys}"
 end
 
 def base_meta_data(result_type = :case)